--- src/index-test.js
@@ -1,1 +1,0 @@
-import './main.js';
