--- src/index.js
@@ -1,4 +1,0 @@
-// import main js
-import './main.js';
-// import css
-import './index.css';
